cleanup, fixes here and there
Cherry-picked from upstream 9a1c0369, clean - no conflicts. The only hunk in a guarded
file is a brace-style reformat in DefaultGrindEngine's repair-npc routine, far from any
customization; all 18 guards are untouched.

Known inherited breakage: upstream's Vector3 rewrite in this commit drops Divide, Limit,
ToArray, RotateRadians, Subtract and the comparison operators without updating
AmeisenBotX.Test/Vector3Tests.cs, so the test project fails with 14 errors in both
UseCustomChanges configurations. Upstream fixes it in the next commit, 1e0ef441. The bot
itself builds clean.

+++ AmeisenBotX.Core/Engines/Quest/Objects/Objectives/CollectQuestObjective.cs
@@ -25,7 +25,7 @@ namespace AmeisenBotX.Core.Engines.Quest.Objects.Objectives
                 }
 
                 IWowInventoryItem inventoryItem = Bot.Character.Inventory.Items.Find(item => item.Id == ItemId);
-                return inventoryItem != null ? Math.Min(100.0 * ((float)inventoryItem.Count) / ((float)WantedItemAmount), 100.0) : 0.0;
+                return inventoryItem != null ? Math.Min(100.0 * inventoryItem.Count / WantedItemAmount, 100.0) : 0.0;
             }
         }
